segment로 분리되지 않은 코드들 분리
