+++ src/entities/tag/index.ts
@@ -1,0 +1,1 @@
+export type { Tag } from "./model/types.ts"
