+++ src/features/postInteraction/comments/index.ts
@@ -1,0 +1,2 @@
+export { PostComments } from "./ui/PostComments"
+export type { PostCommentsProps, NewComment } from "./model/types"
