+++ src/features/postView/index.ts
@@ -1,3 +1,3 @@
 export { PostAuthor } from "./postAuthor/ui/PostAuthor"
-export { PostDetailButton } from "./postDetail/ui/PostDetailButton"
+export { PostDetailButtonAndDialog } from "./postDetail/ui/PostDetailButtonAndDialog.tsx"
 export * from "./postList"
